+++ glamod_marine_processing/obs_suite/scripts/level1e.py
@@ -240,7 +240,18 @@ def process_table(table_df, table_name, pass_time=None):
             return
         table_df.set_index("report_id", inplace=True, drop=False)
 
-    qc_dict[table_name] = {"total": len(table_df)}
+    previous = len(table_df)
+    table_df = table_df[table_df["report_id"].isin(report_ids)]
+    total = len(table_df)
+    removed = previous - total
+    qc_dict[table_name] = {
+        "total": total,
+        "deleted": removed,
+    }
+    if table_df.empty:
+        logging.warning(f"Empty table {table_name}.")
+        return
+
     if flag:
         qc = table_qc.get(table_name).get("qc")
         element = table_qc.get(table_name).get("element")
@@ -290,6 +301,7 @@ def process_table(table_df, table_name, pass_time=None):
     odata_filename = os.path.join(
         params.level_path, FFS.join([table_name, params.fileID]) + ".psv"
     )
+
     table_to_csv(table_df, odata_filename, columns=cdm_columns)
 
 
@@ -430,9 +442,24 @@ for table in obs_tables:
         tables_in.append(table)
 
 if len(tables_in) == 1:
-    logging.warning(
+    logging.error(
         f"NO OBS TABLES AVAILABLE: {params.sid_dck}, period {params.year}-{params.month}"
     )
+    sys.exit()
+
+# Remove report_ids without any observations
+report_ids = pd.Series()
+for table_in in tables_in:
+    df_ = cdm.read_tables(
+        params.prev_level_path,
+        params.prev_fileID,
+        cdm_subset=table_in,
+        na_values="null",
+    )
+    if not df_.empty:
+        report_ids = pd.concat([report_ids, df_["report_id"]], ignore_index=True)
+
+report_ids = report_ids[report_ids.duplicated()]
 
 # DO THE DATA PROCESSING ------------------------------------------------------
 header_df.set_index("report_id", inplace=True, drop=False)
@@ -463,8 +490,14 @@ if qc_avail:
 pass_time = None
 if params.no_qc_suite:
     qc_avail = True
+    # Set report_quality to passed if report_quality is not checked
     qc_df["report_quality"] = header_df["report_quality"]
+    qc_df["report_quality"] = qc_df["report_quality"].mask(
+        qc_df["report_quality"] == "2", "0"
+    )
     pass_time = header_df["report_time_quality"]
+
+qc_df = qc_df[qc_df.index.isin(report_ids)]
 
 # 2. APPLY FLAGS, LOOP THROUGH TABLES -----------------------------------------
 
@@ -485,7 +518,7 @@ for table in obs_tables:
     flag = True if table in tables_in and qc_avail else False
     process_table(table, table, pass_time=pass_time)
 
-# 3 wind QC
+# 3. wind QC
 table_wd = cdm.read_tables(
     params.level_path, params.fileID, cdm_subset=["observations-wd"]
 )